+++ src/gen_model.ml
@@ -486,7 +486,7 @@ let to_model (ast : A.model) : M.model =
     let values = List.map (fun (x : A.lident A.decl_gen) ->
         let typ = Option.get (Option.map ptyp_to_type x.typ) in
         let default = Option.map to_mterm x.default in
-        M.mk_asset_item x.name typ typ ?default:default) a.fields in
+        M.mk_asset_item x.name typ typ ?default:default ~shadow:x.shadow) a.fields in
     let r : M.asset = M.mk_asset a.name (unloc (Option.get a.key)) ~values:values ~sort:(List.map unloc (a.sort)) ~invariants:(List.map (fun x -> to_label_lterm x) a.specs) in
     M.Dasset r
   in